Додано game_screen для бульбашки Скріни переміщені в окремі файли Додано MDCard для налаштувань в main_screen Додано MDTopAppBar для меню в main_screen main_screen зроблений статично
