+++ screens/gamescreen.kv
@@ -1,0 +1,20 @@
+<GameScreen>:
+    id: game_screen
+
+
+<EmdrBall>:
+    size: 50, 50
+    color: 1, 1, 0, 1
+    canvas:
+        Color:
+            rgba: self.color
+        Ellipse:
+            pos: self.pos
+            size: self.size
+
+<EmdrPlace>:
+    ball: pong_ball
+
+    EmdrBall:
+        id: pong_ball
+        center: self.parent.center
